+++ deltafi-ui/src/pages/QueueMetricsPage.vue
@@ -9,7 +9,11 @@
             <template #loading>Loading action queue metrics data. Please wait.</template>
             <Column header="Queue Name" field="name" :sortable="true" />
             <Column header="Queue Size" field="size" :sortable="true" class="metric-column" />
-            <Column header="Timestamp" field="timestamp" :sortable="true" class="metric-column" />
+            <Column header="Timestamp" field="timestamp" :sortable="true" class="metric-column">
+              <template #body="row">
+                <Timestamp :timestamp="row.data.timestamp" />
+              </template>
+            </Column>
           </DataTable>
         </CollapsiblePanel>
       </div>
@@ -24,6 +28,7 @@ import CollapsiblePanel from "@/components/CollapsiblePanel.vue";
 import PageHeader from "@/components/PageHeader.vue";
 import useQueueMetrics from "@/composables/useQueueMetrics";
 import { computed, onMounted, onUnmounted } from "vue";
+import Timestamp from "@/components/Timestamp.vue";
 
 const refreshInterval = 5000; // 5 seconds
 const { data: queueMetrics, loaded, loading, fetch: fetchQueueMetrics } = useQueueMetrics();